kamino_lend

Module cpi

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

Return

Functions

borrow_obligation_liquidity

borrow_obligation_liquidity_v2

delete_referrer_state_and_short_url

deposit_and_withdraw

deposit_obligation_collateral

deposit_obligation_collateral_v2

deposit_reserve_liquidity

deposit_reserve_liquidity_and_obligation_collateral

deposit_reserve_liquidity_and_obligation_collateral_v2

flash_borrow_reserve_liquidity

flash_repay_reserve_liquidity

idl_missing_types

init_farms_for_reserve

init_lending_market

init_obligation

init_obligation_farms_for_reserve

init_referrer_state_and_short_url

init_referrer_token_state

init_reserve

init_user_metadata

liquidate_obligation_and_redeem_reserve_collateral

liquidate_obligation_and_redeem_reserve_collateral_v2

mark_obligation_for_deleveraging

redeem_fees

redeem_reserve_collateral

refresh_obligation

refresh_obligation_farms_for_reserve

refresh_reserve

repay_and_withdraw_and_redeem

repay_obligation_liquidity

repay_obligation_liquidity_v2

request_elevation_group

socialize_loss

socialize_loss_v2

update_lending_market

update_lending_market_owner

update_reserve_config

withdraw_obligation_collateral

withdraw_obligation_collateral_and_redeem_reserve_collateral

withdraw_obligation_collateral_and_redeem_reserve_collateral_v2

withdraw_obligation_collateral_v2

withdraw_protocol_fee

withdraw_referrer_fees