kamino_lend::cpi

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

BorrowObligationLiquidity

Generated CPI struct of the accounts for BorrowObligationLiquidity.

BorrowObligationLiquidityV2

Generated CPI struct of the accounts for BorrowObligationLiquidityV2.

BorrowObligationLiquidityV2BorrowAccounts

Generated CPI struct of the accounts for BorrowObligationLiquidityV2BorrowAccounts.

BorrowObligationLiquidityV2FarmsAccounts

Generated CPI struct of the accounts for BorrowObligationLiquidityV2FarmsAccounts.

DeleteReferrerStateAndShortUrl

Generated CPI struct of the accounts for DeleteReferrerStateAndShortUrl.

DepositAndWithdraw

Generated CPI struct of the accounts for DepositAndWithdraw.

DepositAndWithdrawDepositAccounts

Generated CPI struct of the accounts for DepositAndWithdrawDepositAccounts.

DepositAndWithdrawDepositFarmsAccounts

Generated CPI struct of the accounts for DepositAndWithdrawDepositFarmsAccounts.

DepositAndWithdrawWithdrawAccounts

Generated CPI struct of the accounts for DepositAndWithdrawWithdrawAccounts.

DepositAndWithdrawWithdrawFarmsAccounts

Generated CPI struct of the accounts for DepositAndWithdrawWithdrawFarmsAccounts.

DepositObligationCollateral

Generated CPI struct of the accounts for DepositObligationCollateral.

DepositObligationCollateralV2

Generated CPI struct of the accounts for DepositObligationCollateralV2.

DepositObligationCollateralV2DepositAccounts

Generated CPI struct of the accounts for DepositObligationCollateralV2DepositAccounts.

DepositObligationCollateralV2FarmsAccounts

Generated CPI struct of the accounts for DepositObligationCollateralV2FarmsAccounts.

DepositReserveLiquidity

Generated CPI struct of the accounts for DepositReserveLiquidity.

DepositReserveLiquidityAndObligationCollateral

Generated CPI struct of the accounts for DepositReserveLiquidityAndObligationCollateral.

DepositReserveLiquidityAndObligationCollateralV2

Generated CPI struct of the accounts for DepositReserveLiquidityAndObligationCollateralV2.

DepositReserveLiquidityAndObligationCollateralV2DepositAccounts

Generated CPI struct of the accounts for DepositReserveLiquidityAndObligationCollateralV2DepositAccounts.

DepositReserveLiquidityAndObligationCollateralV2FarmsAccounts

Generated CPI struct of the accounts for DepositReserveLiquidityAndObligationCollateralV2FarmsAccounts.

FlashBorrowReserveLiquidity

Generated CPI struct of the accounts for FlashBorrowReserveLiquidity.

FlashRepayReserveLiquidity

Generated CPI struct of the accounts for FlashRepayReserveLiquidity.

IdlMissingTypes

Generated CPI struct of the accounts for IdlMissingTypes.

InitFarmsForReserve

Generated CPI struct of the accounts for InitFarmsForReserve.

InitLendingMarket

Generated CPI struct of the accounts for InitLendingMarket.

InitObligation

Generated CPI struct of the accounts for InitObligation.

InitObligationFarmsForReserve

Generated CPI struct of the accounts for InitObligationFarmsForReserve.

InitReferrerStateAndShortUrl

Generated CPI struct of the accounts for InitReferrerStateAndShortUrl.

InitReferrerTokenState

Generated CPI struct of the accounts for InitReferrerTokenState.

InitReserve

Generated CPI struct of the accounts for InitReserve.

InitUserMetadata

Generated CPI struct of the accounts for InitUserMetadata.

LiquidateObligationAndRedeemReserveCollateral

Generated CPI struct of the accounts for LiquidateObligationAndRedeemReserveCollateral.

LiquidateObligationAndRedeemReserveCollateralV2

Generated CPI struct of the accounts for LiquidateObligationAndRedeemReserveCollateralV2.

LiquidateObligationAndRedeemReserveCollateralV2CollateralFarmsAccounts

Generated CPI struct of the accounts for LiquidateObligationAndRedeemReserveCollateralV2CollateralFarmsAccounts.

LiquidateObligationAndRedeemReserveCollateralV2DebtFarmsAccounts

Generated CPI struct of the accounts for LiquidateObligationAndRedeemReserveCollateralV2DebtFarmsAccounts.

LiquidateObligationAndRedeemReserveCollateralV2LiquidationAccounts

Generated CPI struct of the accounts for LiquidateObligationAndRedeemReserveCollateralV2LiquidationAccounts.

MarkObligationForDeleveraging

Generated CPI struct of the accounts for MarkObligationForDeleveraging.

RedeemFees

Generated CPI struct of the accounts for RedeemFees.

RedeemReserveCollateral

Generated CPI struct of the accounts for RedeemReserveCollateral.

RefreshObligation

Generated CPI struct of the accounts for RefreshObligation.

RefreshObligationFarmsForReserve

Generated CPI struct of the accounts for RefreshObligationFarmsForReserve.

RefreshObligationFarmsForReserveBaseAccounts

Generated CPI struct of the accounts for RefreshObligationFarmsForReserveBaseAccounts.

RefreshReserve

Generated CPI struct of the accounts for RefreshReserve.

RepayAndWithdrawAndRedeem

Generated CPI struct of the accounts for RepayAndWithdrawAndRedeem.

RepayAndWithdrawAndRedeemCollateralFarmsAccounts

Generated CPI struct of the accounts for RepayAndWithdrawAndRedeemCollateralFarmsAccounts.

RepayAndWithdrawAndRedeemDebtFarmsAccounts

Generated CPI struct of the accounts for RepayAndWithdrawAndRedeemDebtFarmsAccounts.

RepayAndWithdrawAndRedeemRepayAccounts

Generated CPI struct of the accounts for RepayAndWithdrawAndRedeemRepayAccounts.

RepayAndWithdrawAndRedeemWithdrawAccounts

Generated CPI struct of the accounts for RepayAndWithdrawAndRedeemWithdrawAccounts.

RepayObligationLiquidity

Generated CPI struct of the accounts for RepayObligationLiquidity.

RepayObligationLiquidityV2

Generated CPI struct of the accounts for RepayObligationLiquidityV2.

RepayObligationLiquidityV2FarmsAccounts

Generated CPI struct of the accounts for RepayObligationLiquidityV2FarmsAccounts.

RepayObligationLiquidityV2RepayAccounts

Generated CPI struct of the accounts for RepayObligationLiquidityV2RepayAccounts.

RequestElevationGroup

Generated CPI struct of the accounts for RequestElevationGroup.

SocializeLoss

Generated CPI struct of the accounts for SocializeLoss.

SocializeLossV2

Generated CPI struct of the accounts for SocializeLossV2.

SocializeLossV2FarmsAccounts

Generated CPI struct of the accounts for SocializeLossV2FarmsAccounts.

SocializeLossV2SocializeLossAccounts

Generated CPI struct of the accounts for SocializeLossV2SocializeLossAccounts.

UpdateLendingMarket

Generated CPI struct of the accounts for UpdateLendingMarket.

UpdateLendingMarketOwner

Generated CPI struct of the accounts for UpdateLendingMarketOwner.

UpdateReserveConfig

Generated CPI struct of the accounts for UpdateReserveConfig.

WithdrawObligationCollateral

Generated CPI struct of the accounts for WithdrawObligationCollateral.

WithdrawObligationCollateralAndRedeemReserveCollateral

Generated CPI struct of the accounts for WithdrawObligationCollateralAndRedeemReserveCollateral.

WithdrawObligationCollateralAndRedeemReserveCollateralV2

Generated CPI struct of the accounts for WithdrawObligationCollateralAndRedeemReserveCollateralV2.

WithdrawObligationCollateralAndRedeemReserveCollateralV2FarmsAccounts

Generated CPI struct of the accounts for WithdrawObligationCollateralAndRedeemReserveCollateralV2FarmsAccounts.

WithdrawObligationCollateralAndRedeemReserveCollateralV2WithdrawAccounts

Generated CPI struct of the accounts for WithdrawObligationCollateralAndRedeemReserveCollateralV2WithdrawAccounts.

WithdrawObligationCollateralV2

Generated CPI struct of the accounts for WithdrawObligationCollateralV2.

WithdrawObligationCollateralV2FarmsAccounts

Generated CPI struct of the accounts for WithdrawObligationCollateralV2FarmsAccounts.

WithdrawObligationCollateralV2WithdrawAccounts

Generated CPI struct of the accounts for WithdrawObligationCollateralV2WithdrawAccounts.

WithdrawProtocolFee

Generated CPI struct of the accounts for WithdrawProtocolFee.

WithdrawReferrerFees

Generated CPI struct of the accounts for WithdrawReferrerFees.