kamino_lend

Function entry

pub fn entry<'info>(
    program_id: &Pubkey,
    accounts: &'info [AccountInfo<'info>],
    data: &[u8],
) -> ProgramResult

The Anchor codegen exposes a programming model where a user defines a set of methods inside of a #[program] module in a way similar to writing RPC request handlers. The macro then generates a bunch of code wrapping these user defined methods into something that can be executed on Solana.

These methods fall into one category for now.

Global methods - regular methods inside of the #[program].

Care must be taken by the codegen to prevent collisions between methods in these different namespaces. For this reason, Anchor uses a variant of sighash to perform method dispatch, rather than something like a simple enum variant discriminator.

The execution flow of the generated code can be roughly outlined:

Start program via the entrypoint.

Check whether the declared program id matches the input program id. If it’s not, return an error.

Find and invoke the method based on whether the instruction data starts with the method’s discriminator.

Run the method handler wrapper. This wraps the code the user actually wrote, deserializing the accounts, constructing the context, invoking the user’s code, and finally running the exit routine, which typically persists account changes.

The entry function here, defines the standard entry to a Solana program, where execution begins.