kamino_lend

Module instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

BorrowObligationLiquidity

Instruction.

BorrowObligationLiquidityV2

Instruction.

DeleteReferrerStateAndShortUrl

Instruction.

DepositAndWithdraw

Instruction.

DepositObligationCollateral

Instruction.

DepositObligationCollateralV2

Instruction.

DepositReserveLiquidity

Instruction.

DepositReserveLiquidityAndObligationCollateral

Instruction.

DepositReserveLiquidityAndObligationCollateralV2

Instruction.

FlashBorrowReserveLiquidity

Instruction.

FlashRepayReserveLiquidity

Instruction.

IdlMissingTypes

Instruction.

InitFarmsForReserve

Instruction.

InitLendingMarket

Instruction.

InitObligation

Instruction.

InitObligationFarmsForReserve

Instruction.

InitReferrerStateAndShortUrl

Instruction.

InitReferrerTokenState

Instruction.

InitReserve

Instruction.

InitUserMetadata

Instruction.

LiquidateObligationAndRedeemReserveCollateral

Instruction.

LiquidateObligationAndRedeemReserveCollateralV2

Instruction.

MarkObligationForDeleveraging

Instruction.

RedeemFees

Instruction.

RedeemReserveCollateral

Instruction.

RefreshObligation

Instruction.

RefreshObligationFarmsForReserve

Instruction.

RefreshReserve

Instruction.

RepayAndWithdrawAndRedeem

Instruction.

RepayObligationLiquidity

Instruction.

RepayObligationLiquidityV2

Instruction.

RequestElevationGroup

Instruction.

SocializeLoss

Instruction.

SocializeLossV2

Instruction.

UpdateLendingMarket

Instruction.

UpdateLendingMarketOwner

Instruction.

UpdateReserveConfig

Instruction.

WithdrawObligationCollateral

Instruction.

WithdrawObligationCollateralAndRedeemReserveCollateral

Instruction.

WithdrawObligationCollateralAndRedeemReserveCollateralV2

Instruction.

WithdrawObligationCollateralV2

Instruction.

WithdrawProtocolFee

Instruction.

WithdrawReferrerFees

Instruction.